+++ upyog/image/__init__.py
@@ -3,3 +3,4 @@ from .pil_operators import *
 from .draw import *
 from .composition import *
 from .visualiser import *
+from .utils import *