+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Save image with Exif metadata
+
 ## 0.5.5 -- 1 September 2022
 * Add a bunch of utilities related to iterating in `utils.utils`
 * `make_img_grid` in non-verbose mode... finally